fix: add handler to remove empty session on tab close
